+++ app/constants/namespaces.js
@@ -99,5 +99,11 @@ angular.module("orsApp").constant("orsNamespaces", {
     GetCapabilities:
       "https://sgx.geodatenzentrum.de/wms_topplus_web_open?request=GetCapabilities&service=wms",
     url: "https://sgx.geodatenzentrum.de/wms_topplus_web_open?"
+  },
+  layerWorldImagery: {
+    url:
+      "https://server.arcgisonline.com/ArcGIS/rest/services/World_Imagery/MapServer/tile/{z}/{y}/{x}",
+    attribution:
+      "Tiles &copy; Esri &mdash; Source: Esri, i-cubed, USDA, USGS, AEX, GeoEye, Getmapping, Aerogrid, IGN, IGP, UPR-EGP, and the GIS User Community"
   }
 });